% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtokentome.R
\name{bpe_decode}
\alias{bpe_decode}
\title{Decode Byte Pair Encoding sequences to text}
\usage{
bpe_decode(model, x, ...)
}
\arguments{
\item{model}{an object of class \code{youtokentome} as returned by \code{\link{bpe_load_model}}}

\item{x}{an integer vector of BPE id's}

\item{...}{further arguments passed on to youtokentome_encode_as_ids}
}
\description{
Decode a sequence of Byte Pair Encoding ids into text again
}
