% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtokentome.R
\name{bpe_encode}
\alias{bpe_encode}
\title{Tokenise text alongside a Byte Pair Encoding model}
\usage{
bpe_encode(model, x, type = c("subwords", "ids"), bos = FALSE,
  eos = FALSE, reverse = FALSE)
}
\arguments{
\item{model}{an object of class \code{youtokentome} as returned by \code{\link{bpe_load_model}}}

\item{x}{a character vector of text to tokenise}

\item{type}{a character vector of text to tokenise}

\item{bos}{logical if set to TRUE then token 'beginning of sentence' will be added}

\item{eos}{logical if set to TRUE then token 'end of sentence' will be added}

\item{reverse}{logical if set to TRUE the output sequence of tokens will be reversed}
}
\description{
Tokenise text alongside a Byte Pair Encoding model
}
